requirements.txt
astroid==3.2.2
Click==8.1.7
ecdsa==0.19.0
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
future==0.18.3
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
lazy-object-proxy==1.10.0
MarkupSafe==2.1.5
mccabe==0.7.0
pycryptodome==3.3.1
pylint==3.2.3
python-jose-cryptodome==1.3.2
six==1.16.0
typed-ast==1.5.5
Werkzeug==3.0.3
wrapt==1.16.0
Flask-Cors==4.0.1
cryptography==42.0.8
jose==1.0.0
python-jose==3.3.0
psycopg2-binary==2.9.9
